pyproject.toml
[tool.poetry]
name = "nutrition_tracking_api"
version = "0.1.0"
description = ""
authors = []
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
# FastAPI
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
# Database (Sync)
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
# Settings & Config
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
# Logging
loguru = "^0.7.2"
# Metrics
prometheus-client = "^0.19.0"
prometheus-fastapi-instrumentator = "^7.0.0"
sentry-sdk = "^2.52.0"
pyjwt = {extras = ["crypto"], version = "^2.11.0"}
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
factory-boy = "^3.3.0"
httpx = "^0.25.0"
# Code Quality
black = "^23.12.0"
ruff = "^0.8.6"
mypy = "^1.7.0"
# Type stubs
types-python-dateutil = "^2.8.19"
types-psycopg2 = "^2.9.21.20251012"
pytest-lazy-fixtures = "^1.4.0"
greenlet = "^3.3.2"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["prometheus_client.*", "prometheus_fastapi_instrumentator.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=nutrition_tracking_api --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
